% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07. all_data.R
\name{al_data}
\alias{al_data}
\title{df total}
\usage{
al_data(enaho_panel_data, hist_df)
}
\arguments{
\item{enaho_panel_data}{df de enaho}

\item{hist_df}{df de ejecucion}
}
\value{
df conjunto
}
\description{
df total
}
\examples{
al_data(enaho_panel_data,hist_df)
}
