% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.1. binding_index_function.R
\name{binding_index}
\alias{binding_index}
\title{Es una funcion-input para elegir los indices de las variables}
\usage{
binding_index(general_index_modyy, col_facto_modyy, col_index_modyy)
}
\arguments{
\item{general_index_modyy}{indice general de la descarga}

\item{col_facto_modyy}{indice del factor de expansion}

\item{col_index_modyy}{indice del modulo}
}
\value{
un df de indices
}
\description{
Es una funcion-input para elegir los indices de las variables
}
\examples{
binding_index(general_index_modyy,col_facto_modyy,col_index_modyy)
}
