% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04.1. process_&_clean_enaho.R
\name{clean_data}
\alias{clean_data}
\title{Limpieza de datos, corregir factores de expansion y ubigeo}
\usage{
clean_data(enaho_bind_mod03, enaho_bind_mod05, enaho_bind_mod85, caracter)
}
\arguments{
\item{enaho_bind_mod03}{df modulo 3}

\item{enaho_bind_mod05}{df modulo 5}

\item{enaho_bind_mod85}{df modulo 85}
}
\value{
un df para cada modulo limpiado
}
\description{
Limpieza de datos, corregir factores de expansion y ubigeo
}
\examples{
clean_data(enaho_bind_mod03,enaho_bind_mod05,enaho_bind_mod85)
}
