% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02. download_rawdata_enaho.R
\name{download_raw_enaho}
\alias{download_raw_enaho}
\title{descarga los .zip en la carpeta de archivos}
\usage{
download_raw_enaho(all_link)
}
\arguments{
\item{all_link}{requiere los enlaces escrapeados}
}
\value{
retorna los datos descargados para cada modulo
}
\description{
descarga los .zip en la carpeta de archivos
}
\examples{
download_raw_enaho(all_link)
}
