% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.4. build_panel_data_enaho_mods.R
\name{enaho_panel_data}
\alias{enaho_panel_data}
\title{Unificando todos los modulos}
\usage{
enaho_panel_data(enaho_panel_mod03, enaho_panel_mod05, enaho_panel_mod85)
}
\arguments{
\item{enaho_panel_mod03}{modulo 3}

\item{enaho_panel_mod05}{modulo 5}

\item{enaho_panel_mod85}{modulo 85}
}
\value{
retorna un data panel para todas las variables
}
\description{
Unificando todos los modulos
}
\examples{
enaho_panel_data(enaho_panel_mod03,enaho_panel_mod05,enaho_panel_mod85)
}
