% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.1. build_panel_data_enaho_mod03.R
\name{panel_data_03}
\alias{panel_data_03}
\title{creando panel para modulo 3}
\usage{
panel_data_03(enaho_mod03_df)
}
\arguments{
\item{enaho_mod03_df}{df del modulo}
}
\value{
df en data panel
}
\description{
creando panel para modulo 3
}
\examples{
panel_data_03(enaho_mod03_df)
}
