% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.2. build_panel_data_enaho_mod05.R
\name{panel_data_05}
\alias{panel_data_05}
\title{creando data panel para modulo 5}
\usage{
panel_data_05(enaho_mod05_df)
}
\arguments{
\item{enaho_mod05_df}{df del modulo}
}
\value{
df en formato panel
}
\description{
creando data panel para modulo 5
}
\examples{
panel_data_05(enaho_mod05_df)
}
