% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05.3. build_panel_data_enaho_mod85.R
\name{panel_data_85}
\alias{panel_data_85}
\title{creando data panel para modulo 85}
\usage{
panel_data_85(enaho_mod85_df)
}
\arguments{
\item{enaho_mod85_df}{df del modulo}
}
\value{
df en formato panel
}
\description{
creando data panel para modulo 85
}
\examples{
panel_data_85(enaho_mod85_df)
}
