% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.datos_poblacionales.R
\name{poblacion_data}
\alias{poblacion_data}
\title{scraping daos de datos poblacionales}
\usage{
poblacion_data()
}
\value{
df poblacional
}
\description{
scraping daos de datos poblacionales
}
\examples{
poblacion_data()
}
