% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06. datos_presupuestos.R
\name{presupuesto_data}
\alias{presupuesto_data}
\title{scraping daos de ejecucion presupuestal}
\usage{
presupuesto_data(start_date, end_date)
}
\arguments{
\item{start_date}{year de inicio}

\item{end_date}{year final}
}
\value{
df presupuestal
}
\description{
scraping daos de ejecucion presupuestal
}
\examples{
presupuesto_data(start_date,end_date)
}
