% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01. scraping_links_enaho.R
\name{scraping_enaho_datalink_function}
\alias{scraping_enaho_datalink_function}
\title{Scrapea los enlaces de enaho}
\usage{
scraping_enaho_datalink_function(start_date, end_date)
}
\arguments{
\item{start_date}{year de inicio}

\item{end_date}{year final}
}
\value{
df de enlaces de modulo 3, 5, 85
}
\description{
Scrapea los enlaces de enaho
}
\examples{
scraping_enaho_datalink_function(start_date,end_date)
}
