% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.1. select_goals_function.R
\name{select_goals}
\alias{select_goals}
\title{es la 2da pre-funcion}
\usage{
select_goals(doc, var)
}
\arguments{
\item{doc}{el documento o modulo a analizar}

\item{var}{la variable a extraer}
}
\value{
un df para cada modulo de cada variable
}
\description{
es la 2da pre-funcion
}
\examples{
select_goals(doc,var)
}
