% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.1. select_subset_function.R
\name{select_subset}
\alias{select_subset}
\title{elige el sub-conjunto de datos de interes}
\usage{
select_subset(enaho_modrr, global_index_modrr)
}
\arguments{
\item{enaho_modrr}{modulo de analisis}

\item{global_index_modrr}{los indices de interes}
}
\value{
el df de cada modulo
}
\description{
elige el sub-conjunto de datos de interes
}
\examples{
select_subset(enaho_modrr,global_index_modrr)
}
