% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03.2. select_&_biding_enaho.R
\name{selected_enaho_data}
\alias{selected_enaho_data}
\title{Eligiendo las variables de interes}
\usage{
selected_enaho_data(
  enaho_data,
  variables_xlsx,
  select_goals,
  binding_index,
  select_subset
)
}
\arguments{
\item{enaho_data}{datos de enaho}

\item{variables_xlsx}{las claves de las variables en excel}

\item{select_goals}{pre-funcion 1}

\item{binding_index}{pre-funcion 2}

\item{select_subset}{pre-funcion 3}
}
\value{
1 df unido para cada modulo
}
\description{
Eligiendo las variables de interes
}
\examples{
selected_enaho_data(enaho_data,variables_xlsx,select_goals,binding_index,select_subset)
}
