% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07.datos_poblacionales.R
\name{poblacion_data}
\alias{poblacion_data}
\title{Title extrae los datos de poblacion provincial}
\usage{
poblacion_data(start_date, end_date, tildes)
}
\arguments{
\item{start_date}{periodo inicial}

\item{end_date}{periodo final}

\item{tildes}{string de tildes en mayuscula}
}
\value{
el df de poblacion
}
\description{
Title extrae los datos de poblacion provincial
}
\examples{
poblacion_data(start_date,end_date,tildes)
}
